Feature auth and permissions (#8)
* Fix some typos that kept userId from getting added to requests

* Add a custom type to return when a user has successfully signed out

* Add password reset flow

* Make 'reset password' flow send a (dev) email

* Make sure that items are associated with users, don't let users upload items if they aren't signed in

* Handle permission changes from admin page

* Restrict ability to delete items

+++ src/resolvers/Mutation.js
@@ -1,12 +1,25 @@
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
+const { randomBytes } = require('crypto');
+const { promisify } = require('util');
+const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
     async createItem(parent, args, ctx, info) {
         // To do: check if they're logged in
+        if (!ctx.request.userId) {
+            throw new Error('You must be logged in to do that');
+        }
 
         const item = await ctx.db.mutation.createItem({
             data: {
+                // This is how we create a relationship between the item and user
+                user: {
+                    connect: {
+                        id: ctx.request.userId
+                    }
+                },
                 ...args
             }
         }, info);
@@ -33,9 +46,15 @@ const Mutations = {
     async deleteItem(parent, args, ctx, info) {
         const where =  { id: args.id };
         // Find the item
-        const item = await ctx.db.query.item({ where }, `{ id title }`);
+        const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
         // Check if they own that item or have permissions
-        // TODO
+        const ownsItem = item.userid === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(permission => {
+            ['ADMIN', 'ITEMDELETE'].includes(permission)});
+
+        if (!ownsItem && hasPermissions) {
+            throw new Error("You don't have permission to do that!")
+        }
         // if so, delete it
         return ctx.db.mutation.deleteItem({ where }, info)
     },
@@ -53,7 +72,7 @@ const Mutations = {
             }
         }, info);
         // Create the JWT token for them
-        const token = jwt.sign({ userID: user.id }, process.env.APP_SECRET);
+        const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
         // We set the JWT as a cookie on the response
         ctx.response.cookie('token', token, {
             httpOnly: true,
@@ -61,6 +80,127 @@ const Mutations = {
         });
 
         return user;
+    },
+    async signin(parent, { email, password }, ctx, info) {
+        // 1. Check if there is a user with that email
+        const user = await ctx.db.query.user({ where: { email: email}});
+        if (!user) {
+            throw new Error(`No such user found for email ${email}`);
+        }
+        
+        // 2. Check if their password is correct
+        const valid = await bcrypt.compare(password, user.password);
+        if (!valid) {
+            throw new Error('Invalid password');
+        }
+        
+        // 3. generate JWT
+        const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
+        
+        // 4. Set the cookie with the token
+        ctx.response.cookie('token', token, {
+            httpOnly: true,
+            maxAge: 1000 * 60 * 60 * 24 * 365,
+        });
+
+        // 5. Return the user
+        return user;
+    },
+    signout(parent, args, ctx, info) {
+        ctx.response.clearCookie('token');
+        return { message: 'Goodbye!' };
+    },
+    async requestReset(parent, args, ctx, info) {
+        // 1. Check if this is a real user
+        const user = await ctx.db.query.user({ where: { email: args.email } });
+        if (!user) {
+            throw new Error(`No such user found for email ${args.email}`)
+        }
+
+        // 2. Set a reset token and expiry on that user
+        const randomBytesPromisified = promisify(randomBytes);
+        const resetToken = (await randomBytesPromisified(20)).toString('hex');
+        const resetTokenExpiry = Date.now() + 3600000 // 1 hour from request time
+        const res = await ctx.db.mutation.updateUser({
+            where: { email: args.email },
+            data: { resetToken, resetTokenExpiry }
+        });
+        
+        // 3. Email them that reset token
+        const mailRes = await transport.sendMail({
+            from: 'hey@gmail.com',
+            to: user.email,
+            subject: 'Your password reset token',
+            html: makeANiceEmail(`Your password reset token is here!
+                \n\n <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">
+                Click here to reset
+                </a>`)
+        })
+
+        // 4. Return the message
+        return { message: 'Thanks!'}
+    },
+    async resetPassword(parent, args, ctx, info) {
+        // 1. Check if the passwords match
+        if (args.password !== args.confirmPassword) {
+            throw new Error("Your passwords don't match!");
+        }
+        // 2. Check if it's a legit reset token
+        const [user] = await ctx.db.query.users({
+            where: {
+                resetToken: args.resetToken,
+                resetTokenExpiry_gte: Date.now() - 3600000
+            }
+        });
+        if (!user) {
+            throw new Error("This token is either invalid or expired");
+        }
+        // 3. Check if it's expired
+        // 4. Hash their new password
+        const password = await bcrypt.hash(args.password, 10);
+        // 5. Save the new password to the user and remove the old reset token fields
+        const updatedUser = await ctx.db.mutation.updateUser({
+            where: { email: user.email },
+            data: {
+                password,
+                resetToken: null,
+                resetTokenExpiry: null
+            }
+        });
+        // 6. Generate JWT
+        const token = jwt.sign({ userId: updatedUser.id}, process.env.APP_SECRET);
+        // 7. Set the JWT cookie
+        ctx.response.cookie('token', token, {
+            httpOnly: true,
+            maxAge: 1000 * 60 * 60 * 24 * 365
+        })
+        // 8. Return the new user
+        return updatedUser;
+    },
+    async updatePermissions(parent, args, ctx, info) {
+        // 1. Check if they're logged in
+        if (!ctx.request.userId) {
+            throw new Error('You must be logged in');
+        }
+        // 2. Query the current user
+        const currentUser = await ctx.db.query.user({
+            where: {
+                id: ctx.request.userId,
+            }
+        }, info);
+        // 3. Check if they have permission to do this
+        hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+        // 4. Update permissions
+        return ctx.db.mutation.updateUser({
+            data: {
+                permissions: {
+                    set: args.permissions,
+                }
+            },
+            where: {
+                id: args.userId
+            }
+        }, info);
     }
 };
